// src/app/core/api/model/orderResponseDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { InvoiceDto } from './invoiceDto';
import { OrderItemDto } from './orderItemDto';
import { CustomerDto } from './customerDto';


export interface OrderResponseDto { 
    id?: number;
    consecutiveNumber?: number;
    status?: string;
    total?: number;
    discountType?: string;
    discountValue?: number;
    createdAt?: string;
    updatedAt?: string;
    tableId?: number;
    tableNumber?: number;
    branchId?: number;
    userId?: number;
    waiterName?: string;
    customer?: CustomerDto;
    invoice?: InvoiceDto;
    items?: Array<OrderItemDto>;
}